fix(discord): break channel info import cycle
