+++ extensions/discord/src/monitor/message-channel-info-state.ts
@@ -1,4 +1,4 @@
-import type { DiscordChannelInfo } from "./message-channel-info.js";
+import type { DiscordChannelInfo } from "./message-channel-info-types.js";
 
 export const discordChannelInfoCacheState = {
   entries: new Map<string, { value: DiscordChannelInfo | null; expiresAt: number }>(),
